Major Refactoring of reader and helper functions
  - Added orbkit.tools to contain core helper functions
  - Added orbkit.read to contain refactured reader and associated helper functions
  - Added new high_level reader functions with automatic file-type
    detection
  - Started work on automated testing callable via "orbkit test"
  - Test cover most readers
  - Added units module

detCI:

  - Refactured reader functions
  - Added partial reader test for ci readers
