+++ orbkit/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: iso-8859-1 -*-
 '''Imports all important orbkit modules'''
 __all__ = ['grid','options','main','QCinfo','display','run_orbkit','init',
-           'main_read','main_output','rho_compute','rho_compute_no_slice',
+           'read','main_output','rho_compute','rho_compute_no_slice',
            'calc_ao','calc_mo','mo_set','gross_atomic_density',
            'mo_transition_flux_density','get_dipole_moment',
            'atomic_populations'
@@ -15,7 +15,6 @@ from . import grid,options,display,main,atomic_populations
 from .qcinfo import QCinfo
 from .grid import grid_init,get_grid,set_grid
 from .main import run_orbkit,init
-from .read import main_read
 from .output import main_output
 from .core import rho_compute,rho_compute_no_slice
 from .extras import calc_ao,calc_mo,mo_set,gross_atomic_density,\
