--- orbkit/ok.py
@@ -1,11 +1,0 @@
-# -*- coding: iso-8859-1 -*-
-'''Import all important orbkit modules'''
-from orbkit import main
-from orbkit import core
-from orbkit import read
-from orbkit import output
-from orbkit import options
-from orbkit import display
-from orbkit import grid
-from orbkit import extras
-from orbkit.qcinfo import QCinfo
