+++ orbkit/read/__init__.py
@@ -1,13 +1,18 @@
-from orbkit.read.high_level import read
-from orbkit.read.molden import read_molden
-from orbkit.read.gamess import read_gamess
-from orbkit.read.gaussian_fchk import read_gaussian_fchk
-from orbkit.read.gaussian_log import read_gaussian_log
-from orbkit.read.aomix import read_aomix
-from orbkit.read.wfx import read_wfx
-from orbkit.read.wfn import read_wfn
-from orbkit.read.cclib import read_with_cclib
+'''
+Build new ORBKIT io interface
+'''
+
+from .high_level import main_read
+from .molden import read_molden
+from .gamess import read_gamess
+from .gaussian_fchk import read_gaussian_fchk
+from .gaussian_log import read_gaussian_log
+from .aomix import read_aomix
+from .wfx import read_wfx
+from .wfn import read_wfn
+from .cclib import read_with_cclib
+from .tools import mo_select
 
 __all__ = ['read', 'read_molden', 'read_gamess', 'read_gaussian_fchk',
            'read_gaussian_log', 'read_aomix', 'read_wfx', 'read_wfn',
-           'read_with_cclib']
+           'read_with_cclib', 'mo_select']
