+++ orbkit/read/high_level.py
@@ -14,11 +14,11 @@ from orbkit.read.cclib import read_with_cclib
 from orbkit.display import display
 
 readers = {'fchk': read_gaussian_fchk, 'wfx': read_wfx, 'wfn': read_wfn, 
-           'cclib': read_with_cclib, 'molden': read_molden, 'gamess': gaussian_log, 
+           'cclib': read_with_cclib, 'molden': read_molden, 'gamess': read_gaussian_log, 
            'aomix': read_aomix}
 
 
-def read(fname, all_mo=False,spin=None, ignore_molden=False, cclib_parser= None, **kwargs):
+def main_read(fname, all_mo=False, spin=None, itype=None, **kwargs):
   '''
   This is the high-lever interface for the
   orbkit reading routines.
@@ -31,14 +31,6 @@ def read(fname, all_mo=False,spin=None, ignore_molden=False, cclib_parser= None,
     If True, all molecular orbitals are returned.
   spin : {None, 'alpha', or 'beta'}, optional
     If not None, returns exclusively 'alpha' or 'beta' molecular orbitals.
-  cclib_parser : str
-    If a cclib file is read, specifies the cclib.parser.
-  ignore_molden: bool, optional
-    Molden files are problematic for automatic detection 
-    as they might be appended to other types of files.
-    Default bahavior is to first check if the file is a molden file
-    which will lead to other output being disregarded.
-    To ignore the possible presence of molden data use ignore_molden = True.
     
   **Note:**
   
@@ -50,20 +42,15 @@ def read(fname, all_mo=False,spin=None, ignore_molden=False, cclib_parser= None,
     See :ref:`Central Variables` for details.
   '''
   
-  filetype = find_filetype(fname, ignore_molden)
-  
-  if filetype == 'cclib':
-    assert cclib_parser not None, 'For cclib files a parser must be specified'
-  
-  display('Loading data from {0} type file {1}\n'.format(filetype, fname))
-  
-  if filetype =! 'cclib':
-    return readers[filetype](fname, all_mo=all_mo, spin=spin, **kwargs)
-  else:
-    return readers[filetype](fname, cclib_parser=cclib_parser, all_mo=all_mo, spin=spin, **kwargs)
+  if itype is None:
+    itype = find_itype(fname)
+ 
+  display('Loading data from {0} type file {1}\n'.format(itype, fname))
+
+  return readers[itype](fname, all_mo=all_mo, spin=spin, **kwargs)
   
 
-def find_filetype(fname, ignore_molden=False):
+def find_itype(fname):
   '''
   This function is used by the high-level read
   to determine what reader to use.
@@ -74,12 +61,6 @@ def find_filetype(fname, ignore_molden=False):
   
   fname: str
     Specifies the filename for the input file.
-  ignore_molden: bool, optional
-    Molden files are problematic for automatic detection 
-    as they might be appended to other types of files.
-    Default bahavior is to first check if the file is a molden file
-    which will lead to other output being disregarded.
-    To ignore the possible presence of molden data use ignore_molden = True.
     
   **Returns:**
   
@@ -90,7 +71,6 @@ def find_filetype(fname, ignore_molden=False):
     - .fchk
     - .wfx
     - .wfn
-    - .cclib
     
   filetypes determined from magic strings:
     - Molden
@@ -99,28 +79,23 @@ def find_filetype(fname, ignore_molden=False):
     - AOMix
   '''
   
-  extensions = ['fchk', 'wfx', 'wfn', 'cclib']
-  if fname.split('.')[-1] in extensions:
+  extensions = ['fchk', 'wfx', 'wfn']
+  if fname.split('.')[-1].lower() in extensions:
     return fname.split('.')[-1]
 
-  molden_regex = re.compile(r"\[[ ]{,}Molden[ ]+Format[ ]{,}\]")
-  gamess_regex = re.compile(r"GAMESS") #This might be too weak - Can someone who knows Gamess please check?
-  gaussian_regex = re.compile(r"Copyright[,\s\(\)c0-9]+Gaussian\s{,},\s+Inc.")
-  aomix_regex = re.compile(r"\[[ ]{,}AOMix[ ]+Format[ ]{,}\]")
+  molden_regex = re.compile(r"\[[ ]{,}[Mm]olden[ ]+[Ff]ormat[ ]{,}\]")
+  gamess_regex = re.compile(r"[Gg][Aa][Mm][Ee][Ss][Ss]") #This might be too weak - Can someone who knows Gamess please check?
+  gaussian_regex = re.compile(r"[Cc]opyright[,\s\(\)c0-9]+[Gg]aussian\s{,},\s+Inc.")
+  aomix_regex = re.compile(r"\[[ ]{,}[Aa][Oo][Mm]ix[ ]+[Ff]ormat[ ]{,}\]")
   
   regexes = {'molden': molden_regex, 'gamess': gamess_regex, 'gaussian_log': gaussian_regex, 'aomix': aomix_regex}
-  
-  if ignore_molden:
-    filetypes = ['gamess', 'gaussian_log', 'aomix']
-  else:
-    filetypes = ['molden', 'gamess', 'gaussian_log', 'aomix']  
+
+  itypes = ['molden', 'gamess', 'gaussian_log', 'aomix']  
     
   with open(fname, 'r') as fd:
     text = fd.read()
-    for regname in filetypes:
-      if regname != 'molden':
-          text = text[:1e3] #Only Molden files can be appended so the entire file has to be scanned.
-      if regexes['regname'].search(text):
+    for regname in itypes:
+      if regexes[regname].search(text):
         return regname
   raise NotImplementedError('File format not reccognized or reader not implemented!')
