+++ orbkit/test/read/outputs_for_testing/1.wfx
@@ -1,0 +1,292 @@
+   ########################################
+   #                                      #
+   #   WFX format file produced by ORCA   #
+   #                                      #
+   ########################################
+
+
+Warning: this subroutine still in the testing stage. 
+<Title> 
+Name  H2O_orca    Run Type  SinglePoint    Method  HFGTO    Basis Set  Gauss 
+</Title> 
+ 
+<Keywords> 
+  GTO 
+</Keywords> 
+ 
+<Number of Nuclei> 
+  3 
+</Number of Nuclei> 
+ 
+<Number of Primitives> 
+ 27 
+</Number of Primitives> 
+ 
+<Number of Occupied Molecular Orbitals> 
+   5 
+</Number of Occupied Molecular Orbitals> 
+ 
+<Number of Perturbations> 
+ 0 
+</Number of Perturbations> 
+ 
+<Nuclear Names> 
+ O1 
+ H2 
+ H3 
+</Nuclear Names> 
+ 
+<Atomic Numbers> 
+  8 
+  1 
+  1 
+</Atomic Numbers> 
+ 
+<Nuclear Charges> 
+ 8.00000000000000E+00  
+ 1.00000000000000E+00  
+ 1.00000000000000E+00  
+</Nuclear Charges> 
+ 
+<Nuclear Cartesian Coordinates> 
+0.00000000000000E+00 0.00000000000000E+00 -2.22042021570565E-01 
+-2.86342776849333E+00 0.00000000000000E+00 1.76198355498978E+00 
+2.86342776849333E+00 0.00000000000000E+00 1.76198355498978E+00 
+</Nuclear Cartesian Coordinates> 
+ 
+<Net Charge> 
+  0 
+</Net Charge> 
+ 
+<Number of Electrons> 
+ 10 
+</Number of Electrons> 
+ 
+<Number of Alpha Electrons> 
+  5 
+</Number of Alpha Electrons> 
+ 
+<Number of Beta Electrons> 
+  5 
+</Number of Beta Electrons> 
+ 
+<Electronic Spin Multiplicity> 
+  1 
+</Electronic Spin Multiplicity> 
+ 
+<Primitive Centers> 
+  1  1  1  1  1  1  1  1  1  1  1 
+  1  1  1  1  1  1  1  1  2  2  2 
+  2  3  3  3  3
+</Primitive Centers> 
+ 
+<Primitive Types> 
+  1  1  1  1  1  1  1  4  2  3  4 
+  2  3  4  2  3  4  2  3  1  1  1 
+  1  1  1  1  1
+</Primitive Types> 
+ 
+<Primitive Exponents> 
+2.26617677850000E+03 3.40870101910000E+02 7.73631351670000E+01  
+2.14796449400000E+01 6.65894331240000E+00 8.09759756700000E-01  
+2.55307722300000E-01 1.77215043170000E+01 1.77215043170000E+01  
+1.77215043170000E+01 3.86355054400000E+00 3.86355054400000E+00  
+3.86355054400000E+00 1.04809208830000E+00 1.04809208830000E+00  
+1.04809208830000E+00 2.76415444100000E-01 2.76415444100000E-01  
+2.76415444100000E-01 1.30107010000000E+01 1.96225720000000E+00  
+4.44537960000000E-01 1.21949620000000E-01 1.30107010000000E+01  
+1.96225720000000E+00 4.44537960000000E-01 1.21949620000000E-01  
+</Primitive Exponents> 
+ 
+<Molecular Orbital Occupation Numbers> 
+ 2.00000000000000E+00 
+ 2.00000000000000E+00 
+ 2.00000000000000E+00 
+ 2.00000000000000E+00 
+ 2.00000000000000E+00 
+</Molecular Orbital Occupation Numbers> 
+ 
+<Molecular Orbital Energies> 
+ -2.06529440471407E+01 
+ -1.17607391494576E+00 
+ -5.09563645324521E-01 
+ -4.41339273759071E-01 
+ -4.33411457119000E-01 
+</Molecular Orbital Energies> 
+ 
+<Molecular Orbital Spin Types> 
+Alpha 
+Alpha 
+Beta 
+Beta 
+Alpha and Beta 
+</Molecular Orbital Spin Types> 
+ 
+# found spherical MO primitives
+# transferred matrix content
+# done loading matrix, l=0
+# done loading matrix, l=1
+# got trafo matrix
+# transformed spherical to cartesian MO primitives
+
+<Molecular Orbital Primitive Coefficients> 
+<MO Number> 
+   1 
+ </MO Number>
+1.25039252184472E+00
+2.25466637843902E+00
+3.31825745198568E+00
+3.30045256404873E+00
+1.30866769991733E+00
+1.80237043506848E-02
+-2.57748086393678E-03
+3.38546104884753E-03
+0.00000000000000E+00
+0.00000000000000E+00
+2.68405461986633E-03
+0.00000000000000E+00
+0.00000000000000E+00
+1.16898903287838E-03
+0.00000000000000E+00
+0.00000000000000E+00
+-2.47790342974692E-04
+0.00000000000000E+00
+0.00000000000000E+00
+1.29806985679058E-05
+2.20209684152253E-05
+2.50697136924462E-05
+1.67522823841268E-04
+1.29806985675602E-05
+2.20209684146390E-05
+2.50697136917787E-05
+1.67522823841430E-04
+<MO Number> 
+   2 
+ </MO Number>
+3.86910458576671E-01
+6.97664123208486E-01
+1.02677238546591E+00
+1.02126299762470E+00
+4.04942616861212E-01
+-3.82434835082518E-01
+-1.16417238222893E-01
+-1.89176962584087E-03
+0.00000000000000E+00
+0.00000000000000E+00
+-1.49982910176724E-03
+0.00000000000000E+00
+0.00000000000000E+00
+-6.53222091003894E-04
+0.00000000000000E+00
+0.00000000000000E+00
+-3.39493156663822E-03
+0.00000000000000E+00
+0.00000000000000E+00
+-5.47082842291834E-03
+-9.28092885571393E-03
+-1.05658490955304E-02
+-2.39285514053971E-03
+-5.47082842291776E-03
+-9.28092885571295E-03
+-1.05658490955292E-02
+-2.39285514053990E-03
+<MO Number> 
+   3 
+ </MO Number>
+0.00000000000000E+00
+0.00000000000000E+00
+-1.28462361347987E-14
+-1.27773066445161E-14
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+-2.27883430222643E-13
+-1.05285153311237E-14
+-2.17785479584549E+00
+-1.80670096289628E-13
+0.00000000000000E+00
+-1.72664258777310E+00
+-7.86874304153227E-14
+0.00000000000000E+00
+-7.52006398777393E-01
+-1.52093461806625E-14
+0.00000000000000E+00
+-1.36434025862956E-01
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+0.00000000000000E+00
+<MO Number> 
+   4 
+ </MO Number>
+-2.23218099224527E-13
+-4.02499482832885E-13
+-5.92370082377901E-13
+-5.89191581888847E-13
+-2.33621292025255E-13
+2.70828237309726E-13
+8.80151005848830E-14
+-6.31126904428299E-12
+-1.37382134622403E+00
+1.22050585390954E-14
+-5.00368800323185E-12
+-1.08919035782699E+00
+0.00000000000000E+00
+-2.17926131473974E-12
+-4.74376182061470E-01
+0.00000000000000E+00
+-4.56771402353823E-13
+-1.00927896244349E-01
+0.00000000000000E+00
+3.75527728200903E-02
+6.37060031745851E-02
+7.25259321008194E-02
+3.50983535815637E-02
+-3.75527728203487E-02
+-6.37060031750234E-02
+-7.25259321013183E-02
+-3.50983535818085E-02
+<MO Number> 
+   5 
+ </MO Number>
+-5.61000975732908E-02
+-1.01157837731658E-01
+-1.48876903514314E-01
+-1.48078069601687E-01
+-5.87146710922935E-02
+6.78745922468932E-02
+2.23269135430384E-02
+-1.56588896851354E+00
+5.41464436222734E-12
+2.54824559554640E-13
+-1.24146503518841E+00
+4.29282777313796E-12
+2.02029509853870E-13
+-5.40696526849928E-01
+1.86965963721108E-12
+8.79901174845432E-14
+-1.13532910104485E-01
+3.98388952342542E-13
+1.70477369653441E-14
+-3.24500211021810E-02
+-5.50494941413469E-02
+-6.26709521132312E-02
+-3.08027073183708E-02
+-3.24500211018807E-02
+-5.50494941408375E-02
+-6.26709521126513E-02
+-3.08027073180898E-02
+</Molecular Orbital Primitive Coefficients> 
+
+<Energy = T + Vne + Vee + Vnn> 
+0.000000 
+</Energy = T + Vne + Vee + Vnn> 
+
+<Virial Ratio (-V/T)> 
+0.000000 
+</Virial Ratio (-V/T)> 
+
